fix minor bug
